requirements.txt
black==24.1.1
openai==1.11.1
python-dotenv==1.0.1
